% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.ntrans.R
\name{def.calc.ntrans}
\alias{def.calc.ntrans}
\title{NEON Soil Inorganic N Concentrations and Net N Transformation Rates}
\usage{
def.calc.ntrans(kclInt, kclIntBlank, kclExt, soilMoist, dropConditions,
  dropFlagged = FALSE, keepAll = FALSE)
}
\arguments{
\item{kclInt}{A data frame containing soil masses and kcl volumes used in kcl
extractions. Data product table name is ntr_internalLab}

\item{kclIntBlank}{A data frame containing information needed to link kcl
extraction samples to procedural blanks. Data product table name is
ntr_internalLabBlanks}

\item{kclExt}{A data frame containing inorganic N concentrations measured in
kcl extractions and blanks. Data product table name is ntr_externalLab}

\item{soilMoist}{A data frame containing soil moisture values. Data product
table name is sls_soilMoisture}

\item{dropConditions}{An optional list of sampleCondition or dataQF values for which to exclude
ammonium and nitrate concentration measurements}

\item{dropFlagged}{An option to exclude ammonium and nitrate concentration measurements for
samples with external lab quality flags. Defaults to F}

\item{keepAll}{An option to keep all variables and blank info used in calculations. 
Defaults to F, meaning only sample information (not blanks), critical input variables, 
and calculated outputs will be included in the output data frame.}
}
\value{
A data frame of soil inorganic N concentrations in micrograms per gram
  in t-initial and t-final soil samples, as well as net N transformation rates
  for t-final cores. Rows for blank samples are not included when keepAll = F (default).
}
\description{
Calculate soil extractable inorganic nitrogen concentrations and
net N transformation rates for NEON L1 data. It is recommended to use the
neonUtilities package to download data prior to running this function.
}
\examples{
\dontrun{
Load data to R using loadByProduct (neonUtilties)

NTR <- loadByProduct(site = "GUAN", dpID = "DP1.10080.001", 
package = "basic", check.size = F)

out <- def.calc.ntrans(kclInt = NTR$ntr_internalLab, kclIntBlank = NTR$ntr_internalLabBlanks, 
kclExt = NTR$ntr_externalLab, soilMoist = NTR$sls_soilMoisture)

# If data is downloaded to a computer

df1 <- "path/to/data/ntr_internalLab"
df2 <- "path/to/data/ntr_internalLabBlanks"
df3 <- "path/to/data/ntr_ntr_externalLab"
df4 <- "path/to/data/sls_soilMoisture"

out <- def.calc.ntrans(kclInt = df1, kclIntBlank = df2, kclExt =df3, soilMoist = df4,
dropConditions = c("deprecatedMethod", "other"), dropFlagged = T)
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Samantha Weintraub \email{sweintraub@battelleecology.org}
}
\keyword{mineralization,}
\keyword{nitrification}
\keyword{nitrogen,}
\keyword{soil,}
