% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.ntrans.R
\name{def.calc.ntrans}
\alias{def.calc.ntrans}
\title{NEON Soil Inorganic N Concentration and Net N Transformation Rates}
\usage{
def.calc.ntrans(kclInt, kclIntBlank, kclExt, soilMoist, toFilter,
  dropFlagged = FALSE, keepAll = FALSE)
}
\arguments{
\item{kclInt}{A data frame containing soil masses and kcl volumes used in kcl
extractions. Data product table name is ntr_internalLab}

\item{kclIntBlank}{A data frame containing information needed to link kcl
extraction samples to procedural blanks. Data product table name is
ntr_internalLabBlanks}

\item{kclExt}{A data frame containing inorganic N concentrations measured in
kcl extractions and blanks. Data product table name is ntr_externalLab}

\item{soilMoist}{A data frame containing soil moisture values. Data product
table name is sls_soilMoisture}

\item{toFilter}{An optional list of condition values for which to filter out
concentration measurements}

\item{dropFlagged}{An option to filter out concentration measurements for
samples with quality flags, defaults to F}

\item{keepAll}{An option to keep all variables used in nitrogen calculations,
defaults to F}
}
\value{
A dataframe of soil inorganic N concentrations in micrograms per gram
  in t-initial and t-final soil cores, as well as net N transformation rates
  for t-final cores
}
\description{
Calculate soil extractable inorganic nitrogen concentrations and net
transformation rates for NEON L1 data. We recommend using the neonDataStackR
function to stack monthly files prior to using this function.
}
\examples{
\dontrun{
df1 <- "path/to/data/ntr_internalLab"
df2 <- "path/to/data/ntr_internalLabBlanks"
df3 <- "path/to/data/ntr_ntr_externalLab"
df4 <- "path/to/data/sls_soilMoisture"

out <- soilKCl(kclInt = df1, kclIntBlank = df2, kclExt =df3, soilMoist = df4,
filter = T, toFilter = c("thawed & warm"), dropFlagged = T)
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Samantha Weintraub \email{sweintraub@battelleecology.org}
}
\keyword{mineralization,}
\keyword{nitrification}
\keyword{nitrogen}
\keyword{soil,}
